auto-set tags and default journal
When creating new documents, automatically set the tags to whatever the current search tags are, if any. Also when setting the journal, if no journal is selected (i.e. viewing all documents, the default), set the default journal. Add concept of default journal (stored in preferences) to the application, with some guardrails to ensure it always exists. Clean-up the journals page slightly (will be replacing it with a context menu in a future release, so very minor adjustments here).