fix code block doubling pastes
When pasting code into code block, the content was doubling. This was caused by the insertData routine being outside of the multi-line block. But after fixing, I discovered another bug where the normalization routine was putting code_lines outside of blocks, which resulted in the serialized content being removed (because mdast doesnt know about code_lines). Anyways, outright removing the normalization logic seems to fix so... here we are.