re-work persistence to index over files
This large set of changes converts the database from source of truth to a cache over the file system,
where the source of truth is a directory of markdown files + attachments. This returns Chronicles
persistence to an earlier design and should improve its ability to work via file-system cloud-sync,
like iCloud, out of the box.

It introduces a "Sync" operation and numerous supporting changes, along with a few semi-related
improvements or fixes uncovered while hacking away. There remain a few related changes to ensure
the sync experience is smooth; as is the user would need to manually sync when switching devices.