feat: track missing images
- add table to track whether image references in notes are resolveable

Many of my older notes have invalid references, from the historical hacking in chronicles and other note taking applications. Additionally I want to build confidence Chronicles in its current form is not misplacing or failing to import images. This change adds a new table and tracks whether each added image is resolveable. Its done in a naive way and will further degrade sync performance, but this will be easy to optimize in a forthcoming performance sprint once the UX is fully worked out.