fix: duplicate images on paste
- fix a bug where dropping a selection of images (or videos, files) would result in an exponential number of duplicates being created
- refactor file handlers and uploadImage as bytes (ArrayBuffer) rather than base64

On investigation I found I was passing the entire data transfer object on each iteration of files while processing. Unfortunately the routine could not simply modify (pull out) the files it could handle, or mark a subset of them as processed before delegating, nor create a new data transfer object with the subset it intended to pass on (delegate) - for browser security reason. Instead I unified all media handling under one hood; this should also make it easier to later handle uploads in the background with progress indicators to user and better error handling. Will defer to a future improvement.