feat: add dark mode
- add dark mode theme
- Refactor many styles to align closer to themed approach (hover:bg-accent)
- Add Theme synchronization hooks
- Refactor Preferences to be an observable store, setup at launch, available on application state

The addition of dark mode is motivated in part as a driver to get Chronicles more aligned with themeing in general. It is also motivated by the app being nearly feature complete for a base notes application, but being subjectively ugly to me personally. I hope this not only makes it more satisfying to use but also easier to build achive more of the stylistic goals I have for Chronicles.