chore: update sqlite, rebuild. pin yargs.
- update sqlite and electron rebuild
- pin yargs

I'm on node v22. This requires a newer version of better-sqlite3, which results in node-gyp issues (missing distutils). Can work around that by updating electron-rebuild, but then that has an esm bug (or...?) in how it imports yargs, or depends on a newer version. So pinning that version to satisfy it. Will address that in a follow up. Also, apparently newer node versions include sqlite3, and maybe electron will pick that up too. So maybe all these hacks and helper packages can be dropped at some point.