fix: address multiple sync issues
- correctly run sync against new directory when changing notes directory
- reactively update search on sync completion (avoids stale document results)
- add new sync store scaffolding
- add sync documentation

When sync ran, it did not update search results. Further when changing (setting) notes directory, sync would run but results would not be reflected in UI until an application refresh. Moving sync to a UI store model simplifies the process of running and watching for changes. Lastly, add documentation of the existing sync process